More aggressive scaling of the interval
Based on the docs for Skritter, adjust the interval more aggressively.
First, when lengthening for a correct answer, use the actual interval
taken, rather than the desired one.  The argument here is that if the
user remembered it for the longer interval, they will continue to
remember it.

Second, use larger factors to change the interval.