Remove use of actual interval
Although using the actual interval probably makes sense for some kind
of learning, it greatly reduces the repetition seen.  It seems to be
better to keep repeating words a number of times, even if that
repeating gets spread out over time.

In other words, if the interval was 5 minutes, the new interval will
based on that 5 minute interval (say 2.2 * 5 minutes, or 0.25 time 5
minutes), rather than the actual amount of time before the word was
written.