timelearn: Create initial timelearn crate
The timelearn crate manages the problem storage for a time-based
learning system.  It is agnostic to the meaning of the questions and
answers (just that they are valid UTF-8), but manages time intervals.
The data is stored in an sqlite3 database.

There is a test that exercises the API, but doesn't perform testing that
the operation is correct.