timelearn: Document most of the crate
Summary:
Create documentation for the public parts of the crate.

timelearn: Forbid missing_docs

Now that everything is documented, change this into an error so that we
get a compilation error on anything added that isn't documented.

Test Plan:
Run `cargo doc` and view the resulting documentation in a browser.
Second change adds a lint to enforce every public entity have a
documentation comment.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D1