refactor: Term -> Steno
Summary:
To prepare to generalize the interaction, rename the Term type to Steno,
and change instances of this to 'user' in the learning code.  This
forces the existing field 'user', which holds the user input strokes to
become 'input' to better represent what was input.

Test Plan:
Refactor, so make sure compilation, lints, and tests all continue to
pass as before.  Run existing code briefly to make sure it still
works.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D4