timelearn: Add `kind` to Store
Summary:
Add a simple string `kind` to the Store.  This indicates how the
problems should be interpreted.  The legacy tables will be created with
"steno" as the kind.

Ensure that Store is a "steno" store

Verify the kind upon opening the store, to ensure we are trying to ask
the only kind we currently support.

Test Plan:
Create a new words.db by running with the create argument and no file
present.  Verify that this runs, and that by examiniation of the
database, the new fields are present.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D8