Implement "simple" User type
Summary:
The simple type prompts the user with the question, waits for a space,
presents the answer, and asks the user to press 1-4 to indicate how well
it was answered.  This is useful for problem types that don't have
implementations, or that require external interaction.

Test Plan:
Copy the steno database and change the type to "simple", and perform
some interactions to make sure that it seems to do the right thing.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D10