Move `Status` into timelearn crate
Summary:
It doesn't really belong here, but in order to be able to implement
learning methods as separate crates, it needs to be available, as well
as Problem.  Possibly, in the future, consider moving Problem and Status
into another crate.

Test Plan: Verify the code compiles and tests still pass

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D13