midilearn: First implementation
Summary:
The MIDI learner (midilearn) listens over MIDI for the user to play the
given exercise.  This implements support for scale practice.  The scales
are described by the intervals, and some parameters about those scales,
and how they should be played.  Currently, support is only for single
hand played scales.

Test Plan:
Modify a problem set by hand to have the "midi" style, and use the
generator to generate data.  Verify that the basics of scale practice
works.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D18