gen: Auto-delete table when updating
Summary:
The gen scripts generate a series of SQL that will update the problem
set in the database.  Add a "DELETE FROM ..." line to this output, to
remove the existing rows before re-adding them.  We also clean up the
major-7 names to use the Unicode 'Δ' symbol instead of a 'M', to make
the exercises easier to read.

Test Plan: View output for correctness.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D22