midilearn: Add 2-handed scales
Summary:
Change the notes involved in a scale to a Vec<Note> to allow an exercise
to be played with both hands.  This will be built in octaves for the
given number of hands (which means the two hands must start one octave
apart, not more).

Test Plan: Verify that the example works with C, and two hands.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Maniphest Tasks: T1

Differential Revision: https://davidb.phacility.com/D25