gen: Fix non initial chord sequences
Summary:
The chords were incorrect for the cases where the sequence didn't start
with the first chord.  Instead of trying to generate the later chords,
which will result in the wrong chord because of the initial interval,
generate the whole sequence, and then just select the appropriate
subsequence.

Test Plan: Look through the output to check that the sequences are now correct.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D32