Show next problem when appropriate
Summary:
Using the new 'next' parameter to the User code, collect two possible
problems.  This will show these on the performance type of solvers,
giving the user a bit of a heads up as to what will be asked next.  The
next problem should be correct, as long as it is shown, since if the
the current problem can at best be asked with 5 seconds, and if
something is shown, it means it is due.

This also gives a little bit of a heads up as to what might be coming up
next to learn.

Test Plan:
Verify the performance of voicing drills, to see that the appropriate
next word is shown.  By playing the current one correctly, the one
that was the "next" one, should then be the one asked.  Also verify
that steno learning still works, and doesn't have modified behavior.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D39