timelearn: Return a new problem if nothing is ready
Summary:
The change to return possibly multiple problems ends up returning a
0-element vector in the case when there are no pending problems.  Fix
this by detecting this situation, and querying for a new problem.

Annotate the new problem with (" NEW") after the question to make it
distinct to the user.

Test Plan:
Test by running the voicing exercise and making sure that new problems
are asked correctly.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D40