gen: Right-hand only for modes
Summary:
Speed up the learning of the modes and additional scales by only doing
them with the right hand.  We could come back later if that was deemed
important or useful.

Test Plan:
Study the output of the script and compare with existing database
entries.  Make sure that they match up until the point where the left
hand isn't included.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D41