Add support for learning MIDI licks
Summary:
The new lick type works like "chords", except that the stopping
condition is more like scales (wrong starting note).  Adds a first lick
from Jaki Byard.

Test Plan:
Generate a lick database, and test a few with the midi keyboard.  The
concern is with the threshold for simultaneous notes.

Reviewers: d3zd3z

Reviewed By: d3zd3z

Differential Revision: https://davidb.phacility.com/D42